backend/pyproject.toml
[tool.poetry]
name = "silent-circle-backend"
version = "0.1.0"
description = "Silent Circle backend: FastAPI + aiogram + PostgreSQL + Redis"
authors = ["BloodCharry <vladvorobei1994@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"

# Web framework
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }

# Telegram bot
aiogram = "^3.10.0"

# Database
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.30.0"

# Redis / background tasks
redis = "^5.0.4"
celery = "^5.4.0"
apscheduler = "^3.10.4"

# Data validation
pydantic = "^2.8.0"
pydantic-settings = "^2.3.0"

# Security
python-jose = "^3.3.0"
passlib = "^1.7.4"

# Utils
httpx = "^0.27.0"
loguru = "^0.7.2"
setuptools = "^80.9.0"
testcontainers = "^4.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.0"
mypy = "^1.10.0"
aiosqlite = "^0.21.0"
psycopg2-binary = "^2.9.11"
pytest-dotenv = "^0.5.2"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"